Add solution for day 3 part 1
In this commit, we are adding a new package for day 3 and solving the part one of this code challenge with three classes.

The WireVector is responsible for giving a wire a direction and a magnitude. A Wire is composed of a list of WireVector's, and the Wire can find its intersections with another Wire.

The CentralPort is composed of a list of Wire's, and it can find the distance to the closest wire intersection.