Add solution for day 3 part 2
In this commit, we are solving the 2nd part of the 3rd day.

The CentralPort class got a new method to sum the distance of each wire to the first intersection between them.

The Wire class implements a new method to calculate the distance mentioned above.

We also refactored the constructor of CentralPort to receive two wires, instead of a list of wires.